worker/requirements.txt
beautifulsoup4==4.10.0
Flask==2.0.1
pika==1.2.0
pymongo==4.0.2
python-dotenv==0.19.2
dnspython==1.16.0
